==> 06_Imdb_Score_Model/05.Ter Cleaning Dataset.ipynb <==
import pandas as pd
import numpy as np
import plotly.figure_factory as ff
import plotly.express as px

dataset = pd.read_csv(f'src/temps/datasetWithPonderation.csv', delimiter=',', on_bad_lines='skip')
print(len(dataset))
'''
for i in range(2,12):
    importDF = pd.read_csv(f'temps/t{i}.csv', delimiter=',', on_bad_lines='skip')
    dataset = dataset.append(importDF)
    print(len(dataset))
'''
dataset = dataset.rename(columns={"originalTitle": "title", "startYear": "year", "primaryName" : "director", "parental_advisor" : "parentalAdvisor", "averageRating" : "imdbRating"})
dataset = dataset.drop(columns=['title', 'director', 'resume'])
dataset = dataset[dataset['genre'] != np.nan ]
#dataset = dataset.dropna(subset=['year'])
#dataset = dataset.dropna(subset=['imdbRating'])
dataset = dataset.dropna(subset=['parentalAdvisor'])
dataset = dataset.dropna(subset=['duree'])
#dataset = dataset.dropna(subset=['genre'])
dataset = dataset[dataset['themaScore'] != 0 ]
dataset = dataset.reindex()
print(len(dataset))
#print(dataset.columns)

dataset = dataset.sort_values(by=['themaScore'], ascending=False)
dataset = dataset.drop_duplicates(subset=['tconst'])

print(len(dataset))
display(dataset.head())

target_name = 'imdbRating'
print(len(dataset))
# ReIndex the dataset for put the TARGET on last column. 
dataset = dataset.reindex(columns = [col for col in dataset.columns if col != target_name] + [target_name])

print('Taille Dataset :')
print(len(dataset)) 
print('Valeurs manquantes en % :')
100*dataset.isnull().sum()/dataset.shape[0]

dataset.to_csv(r'src/dataset.csv', index = False, header=True)

dataset = pd.read_csv(r'src/dataset.csv', delimiter=',', on_bad_lines='skip')

corr_matrix = round(dataset.corr(), 4)

fig = ff.create_annotated_heatmap(corr_matrix.values,
                                  x = corr_matrix.columns.values.tolist(),
                                  y = corr_matrix.index.values.tolist(),
                                  )

fig.show()

###################
###### TEST OUTLIER 
print(len(dataset))

dfnumVotes = dataset.sort_values(by=['numVotes'])
#dfnumVotes = dfnumVotes[dfnumVotes['numVotes'] < 1670000] # 1670000 -> 0.2172
#dfnumVotes = dfnumVotes[dfnumVotes['numVotes'] > 170000] #  30_000 -> 0.3598, 50_000 -> 0.3972, 80_000 -> 0.436, 100_000 -> 0.4547, 160_000 -> 0.5077

dfnumVotes = dfnumVotes[dfnumVotes['numVotes'] < 20000]
dfnumVotes = dfnumVotes[dfnumVotes['numVotes'] > 200]

print(len(dfnumVotes))

px.histogram(dfnumVotes["numVotes"])

corr_matrix = round(dfnumVotes.corr(), 4)

fig = ff.create_annotated_heatmap(corr_matrix.values,
                                  x = corr_matrix.columns.values.tolist(),
                                  y = corr_matrix.index.values.tolist(),
                                  )

fig.show()

dfduree = dfnumVotes.sort_values(by=['duree'])
dfduree = dfduree[dfduree['duree'] < 150]
dfduree = dfduree[dfduree['duree'] > 40]

print(len(dfduree))

px.histogram(dfduree["duree"])

corr_matrix = round(dfduree.corr(), 4)

fig = ff.create_annotated_heatmap(corr_matrix.values,
                                  x = corr_matrix.columns.values.tolist(),
                                  y = corr_matrix.index.values.tolist(),
                                  )


fig.show()

dfyear = dfduree.sort_values(by=['year'])
dfyear = dfyear[dfyear['year'] < 2017]
dfyear = dfyear[dfyear['year'] > 1930]

print(len(dfyear))

px.histogram(dfyear["year"])

corr_matrix = round(dfyear.corr(), 4)

fig = ff.create_annotated_heatmap(corr_matrix.values,
                                  x = corr_matrix.columns.values.tolist(),
                                  y = corr_matrix.index.values.tolist(),
                                  )


fig.show()

dfgenre = dfyear
countGenre = dfgenre['genre'].value_counts()
countGenre = countGenre[countGenre[:] > 400 ] # Drop Less than 5% of biggest genre.
listGenre = countGenre.keys().tolist()

dfgenre = dfgenre[dfgenre['genre'].isin(listGenre)]

print(len(dfgenre))

countGenre.plot(kind='bar')

dfparentalAdvisor = dfgenre
countAdvisor = dfparentalAdvisor['parentalAdvisor'].value_counts()

countAdvisor = countAdvisor[countAdvisor[:] > 1000 ] # Drop Less than 5% of biggest genre.
print(countAdvisor)
listGenre = countAdvisor.keys().tolist()

dfparentalAdvisor = dfparentalAdvisor[dfparentalAdvisor['parentalAdvisor'].isin(listGenre)]

print(len(dfparentalAdvisor))

countAdvisor.plot(kind='bar')

dfimdbRating = dfparentalAdvisor.sort_values(by=['imdbRating'])

print(len(dfimdbRating))

px.histogram(dfimdbRating["imdbRating"])

corr_matrix = round(dfimdbRating.corr(), 4)

fig = ff.create_annotated_heatmap(corr_matrix.values,
                                  x = corr_matrix.columns.values.tolist(),
                                  y = corr_matrix.index.values.tolist(),
                                  )

fig.show()

dfimdbRating.to_csv(r'src/dataset.csv', index = False, header=True)
